Enhancement: Extract BetweenMinimumAndMaximumCountResolutionStrategy
