+++ src/FieldResolution/DefaultStrategy.php
@@ -13,7 +13,6 @@ declare(strict_types=1);
 
 namespace Ergebnis\FactoryBot\FieldResolution;
 
-use Ergebnis\FactoryBot\Count;
 use Ergebnis\FactoryBot\FieldDefinition;
 use Ergebnis\FactoryBot\FixtureFactory;
 use Faker\Generator;
@@ -21,7 +20,7 @@ use Faker\Generator;
 /**
  * @internal
  */
-final class DefaultStrategy implements CountResolutionStrategy, FieldValueResolutionStrategy
+final class DefaultStrategy implements FieldValueResolutionStrategy
 {
     public function resolveFieldValue(
         Generator $faker,
@@ -35,16 +34,6 @@ final class DefaultStrategy implements CountResolutionStrategy, FieldValueResolu
         return $fieldDefinition->resolve(
             $faker,
             $fixtureFactory,
-        );
-    }
-
-    public function resolveCount(
-        Generator $faker,
-        Count $count,
-    ): int {
-        return $faker->numberBetween(
-            $count->minimum(),
-            $count->maximum(),
         );
     }
 }
