+++ test/Unit/FieldResolution/DefaultStrategyTest.php
@@ -130,74 +130,4 @@ final class DefaultStrategyTest extends Test\Unit\AbstractTestCase
 
         self::assertSame($expected, $resolved);
     }
-
-    #[Framework\Attributes\DataProviderExternal(Test\DataProvider\IntProvider::class, 'greaterThanOrEqualToZero')]
-    public function testResolveCountResolvesCountWithFakerWhenCountIsExact(int $value): void
-    {
-        $strategy = new FieldResolution\DefaultStrategy();
-
-        $resolved = $strategy->resolveCount(
-            self::faker(),
-            Count::exact($value),
-        );
-
-        self::assertSame($value, $resolved);
-    }
-
-    #[Framework\Attributes\DataProviderExternal(Test\DataProvider\IntProvider::class, 'greaterThanOrEqualToZero')]
-    public function testResolveCountResolvesCountWithFakerWhenCountIsBetweenAndFakerReturnsMinimum(int $minimum): void
-    {
-        $maximum = self::faker()->numberBetween($minimum + 1);
-
-        $strategy = new FieldResolution\DefaultStrategy();
-
-        $resolved = $strategy->resolveCount(
-            new Test\Double\Faker\MinimumGenerator(),
-            Count::between(
-                $minimum,
-                $maximum,
-            ),
-        );
-
-        self::assertSame($minimum, $resolved);
-    }
-
-    #[Framework\Attributes\DataProviderExternal(Test\DataProvider\IntProvider::class, 'greaterThanZero')]
-    public function testResolveCountResolvesCountWithFakerWhenCountIsBetweenAndFakerReturnsMaximum(int $maximum): void
-    {
-        $minimum = self::faker()->numberBetween(0, $maximum - 1);
-
-        $strategy = new FieldResolution\DefaultStrategy();
-
-        $resolved = $strategy->resolveCount(
-            new Test\Double\Faker\MaximumGenerator(),
-            Count::between(
-                $minimum,
-                $maximum,
-            ),
-        );
-
-        self::assertSame($maximum, $resolved);
-    }
-
-    public function testResolveCountResolvesCountWithFakerWhenCountIsBetween(): void
-    {
-        $faker = self::faker();
-
-        $minimum = $faker->numberBetween(1);
-        $maximum = $faker->numberBetween($minimum + 1);
-
-        $strategy = new FieldResolution\DefaultStrategy();
-
-        $resolved = $strategy->resolveCount(
-            $faker,
-            Count::between(
-                $minimum,
-                $maximum,
-            ),
-        );
-
-        self::assertGreaterThanOrEqual($minimum, $resolved);
-        self::assertLessThanOrEqual($maximum, $resolved);
-    }
 }
